Rename wizard steps accordingly
