--- src/modules/dashboard/components/CreateColonyWizard/StepColonyENSName.css
@@ -1,15 +1,0 @@
-.main {
-  padding-top: 155px;
-  width: 100%;
-  max-width: 460px;
-}
-
-.nameForm {
-  margin-top: 40px;
-}
-
-.buttons {
-  display: flex;
-  justify-content: flex-end;
-  margin-top: 48px;
-}
