+++ src/modules/dashboard/components/CreateColonyWizard/StepColonyName.css
